update todo-mvc for beta2
